Allow for no history

+++ imgt/track_imgt_ref.py
@@ -39,7 +39,7 @@ def main(argv):
 
     # find the latest downloaded file
     file_list = glob.glob(config['file_dir'] + '/' + config['file_prefix'] + '*')
-    oldfile = max(file_list, key=os.path.getctime)
+    oldfile = max(file_list, key=os.path.getctime) if len(file_list) > 0 else None
 
     # download a new reference set from IMGT
     newfile = config['file_dir'] + '/' + config['file_prefix'] + datetime.date.today().strftime('_%Y_%m_%d') + '.fasta'
@@ -47,44 +47,45 @@ def main(argv):
         with open(newfile, 'wb') as fo:
             fo.write(response.read())
 
-    species = config['species']
-    old_recs = read_reference(oldfile, species)
-    new_recs = read_reference(newfile, species)
+    if oldfile:
+        species = config['species']
+        old_recs = read_reference(oldfile, species)
+        new_recs = read_reference(newfile, species)
 
-    changes = False
+        changes = False
 
-    with open(config['log_file'], 'a') as fl:
-        for sp in species:
-            for rec in old_recs[sp].keys():
-                if rec not in new_recs[sp]:
-                    fl.write('%s DROPPED %s %s\n' % (datetime.datetime.now().isoformat(timespec='minutes'), sp, rec))
-                    changes = True
-                elif old_recs[sp][rec] != new_recs[sp][rec]:
-                    fl.write('%s CHANGED %s %s\n' % (datetime.datetime.now().isoformat(timespec='minutes'), sp, rec))
-                    fl.write('old: %s\nnew: %s\n' % (old_recs[sp][rec], new_recs[sp][rec]))
-                    changes = True
-            for rec in new_recs[sp].keys():
-                if rec not in old_recs[sp]:
-                    fl.write('%s ADDED %s %s\n' % (datetime.datetime.now().isoformat(timespec='minutes'), sp, rec))
-                    changes = True
+        with open(config['log_file'], 'a') as fl:
+            for sp in species:
+                for rec in old_recs[sp].keys():
+                    if rec not in new_recs[sp]:
+                        fl.write('%s DROPPED %s %s\n' % (datetime.datetime.now().isoformat(timespec='minutes'), sp, rec))
+                        changes = True
+                    elif old_recs[sp][rec] != new_recs[sp][rec]:
+                        fl.write('%s CHANGED %s %s\n' % (datetime.datetime.now().isoformat(timespec='minutes'), sp, rec))
+                        fl.write('old: %s\nnew: %s\n' % (old_recs[sp][rec], new_recs[sp][rec]))
+                        changes = True
+                for rec in new_recs[sp].keys():
+                    if rec not in old_recs[sp]:
+                        fl.write('%s ADDED %s %s\n' % (datetime.datetime.now().isoformat(timespec='minutes'), sp, rec))
+                        changes = True
 
-        if changes:
-            with open(config['ogre_file'], 'w') as fo, open(newfile, 'r') as fi:
-                fo.write(fi.read())
-            if config['ogre_touch'] != 'none':
-                os.system('touch %s' % config['ogre_touch'])
-            fl.write('%s OGRDB file %s updated\n' % (datetime.datetime.now().isoformat(timespec='minutes'), config['ogre_file']))
+            if changes:
+                with open(config['ogre_file'], 'w') as fo, open(newfile, 'r') as fi:
+                    fo.write(fi.read())
+                if config['ogre_touch'] != 'none':
+                    os.system('touch %s' % config['ogre_touch'])
+                fl.write('%s OGRDB file %s updated\n' % (datetime.datetime.now().isoformat(timespec='minutes'), config['ogre_file']))
 
-            # download a new gapped reference set from IMGT
-            with urllib.request.urlopen(config['gapped_file_url']) as response:
-                with open(config['gapped_ogre_file'], 'wb') as fo:
-                    fo.write(response.read())
-            fl.write('%s OGRDB file %s updated\n' % (datetime.datetime.now().isoformat(timespec='minutes'), config['gapped_ogre_file']))
+                # download a new gapped reference set from IMGT
+                with urllib.request.urlopen(config['gapped_file_url']) as response:
+                    with open(config['gapped_ogre_file'], 'wb') as fo:
+                        fo.write(response.read())
+                fl.write('%s OGRDB file %s updated\n' % (datetime.datetime.now().isoformat(timespec='minutes'), config['gapped_ogre_file']))
 
 
-        else:
-            fl.write('%s No changes detected.\n' % (datetime.datetime.now().isoformat(timespec='minutes')))
-            os.remove(newfile)
+            else:
+                fl.write('%s No changes detected.\n' % (datetime.datetime.now().isoformat(timespec='minutes')))
+                os.remove(newfile)
 
 if __name__ == "__main__":
     main(sys.argv)
